Corrected logs and moved idrac specific code to drac utils
