+++ ironic/drivers/modules/redfish/utils.py
@@ -279,97 +279,3 @@ def get_system(node):
                       {'address': driver_info['address'],
                        'node': node.uuid, 'error': e})
 
-def _ping_host(host):
-    response = subprocess.call(
-        "ping -c 1 {} 2>&1 1>/dev/null".format(host), shell=True)
-    return (response == 0)
-
-def wait_for_host_state(host,
-                        alive=True,
-                        ping_count=3,
-                        retries=24):
-    if alive:
-        ping_type = "pingable"
-
-    else:
-        ping_type = "not pingable"
-
-    LOG.info("Waiting for the iDRAC %s to become %s" % (host,ping_type))
-
-    response_count = 0
-    state_reached = False
-
-    while retries > 0 and not state_reached:
-        response = _ping_host(host)
-        retries -= 1
-        if response == alive:
-            response_count += 1
-            LOG.debug("The iDRAC is %s, count=%s",
-                      ping_type,
-                      response_count)
-            if response_count == ping_count:
-                LOG.debug("Reached specified ping count")
-                state_reached = True
-        else:
-            response_count = 0
-            if alive:
-                LOG.debug("The iDRAC for node %s is still not pingable" % host)
-            else:
-                LOG.debug("The iDRAC for node %s is still pingable" % host)
-        time.sleep(10)
-
-    return state_reached
-
-def wait_for_host(host):
-    LOG.debug("iDRAC was reset, waiting for return to operational state")
-    state_reached = wait_for_host_state(
-                            host,
-                            alive=False,
-                            ping_count=2,
-                            retries=24)
-
-    if not state_reached:
-        raise exceptions.DRACOperationFailed(
-            drac_messages="Timed out waiting for the %s iDRAC to become "
-            "not pingable" % host)
-
-    LOG.info("The iDRAC %s has become not pingable" % host)
-
-    state_reached = wait_for_host_state(
-        host,
-        alive=True,
-        ping_count=3,
-        retries=24)
-    if not state_reached:
-        raise exceptions.DRACOperationFailed(
-            drac_messages="Timed out waiting for the %s iDRAC to become "
-            "pingable" % host)
-
-    LOG.info("The iDRAC %s has become pingable" % host)
-
-def wait_until_idrac_is_ready(oem_manager, host, retries, retry_delay):
-    """Waits until the iDRAC is in a ready state
-    :param retries: The number of times to check if the iDRAC is
-                    ready.
-    :param retry_delay: The number of seconds to wait between
-                        retries.
-    """
-
-    while retries > 0:
-        LOG.debug("Checking to see if the iDRAC %s is ready" % host)
-
-        if oem_manager.is_idrac_ready():
-            LOG.debug("The iDRAC %s is ready" % host)
-            return
-
-        LOG.debug("The iDRAC %s is not ready" % host)
-        retries -= 1
-        if retries > 0:
-            time.sleep(retry_delay)
-
-    if retries == 0:
-        err_msg = ("Timed out waiting for the iDRAC %s "
-                    "to become ready after reset" %
-                    host)
-        LOG.error(err_msg)
-        raise exceptions.DRACOperationFailed(drac_messages=err_msg)
